Fix issue with loading images larger than 4B voxels (#114)
* Change ImageParams to use size_t for image dimensions and adjust everything that comes out of it

* Fix remaining conversino issues in Unit tests

* Fix order of operations to avoid overflow error

* Fix other order of operations error

* Bump version

* Use ssize_t for image sizes

* Add forgotten usages of ssize_t

* Fixes to the Distance-driven kernels

* Forgotten ssize_t usages in the Siddon projector

* Use 64 bits to compute the number of voxels while reading file

* Other fixes related to 32-bin signed integers

* Fix overflow issue when using reinterpretAndCast

* Fix issue with Array.hpp using floats to compute strides

* Fix LaTeX error

+++ yrt-pet/include/yrt-pet/datastruct/image/ImageUtils.cuh
@@ -12,11 +12,10 @@
 namespace yrt::util
 {
 
-inline HOST_DEVICE_CALLABLE void
-    trilinearInterpolateCore(float pos_x, float pos_y, float pos_z, int nx,
-                             int ny, int nz, float origin_x, float origin_y,
-                             float origin_z, float inv_vx, float inv_vy,
-                             float inv_vz, int indices[8], float weights[8])
+inline HOST_DEVICE_CALLABLE void trilinearInterpolateCore(
+    float pos_x, float pos_y, float pos_z, ssize_t nx, ssize_t ny, ssize_t nz,
+    float origin_x, float origin_y, float origin_z, float inv_vx, float inv_vy,
+    float inv_vz, ssize_t indices[8], float weights[8])
 {
 
 	// Convert physical position to grid coordinates
@@ -71,10 +70,10 @@ inline HOST_DEVICE_CALLABLE void
 }
 
 inline HOST_DEVICE_CALLABLE void
-    trilinearInterpolate(float pos_x, float pos_y, float pos_z, int nx, int ny,
-                         int nz, float length_x, float length_y, float length_z,
-                         float off_x, float off_y, float off_z, int indices[8],
-                         float weights[8])
+    trilinearInterpolate(float pos_x, float pos_y, float pos_z, ssize_t nx,
+                         ssize_t ny, ssize_t nz, float length_x, float length_y,
+                         float length_z, float off_x, float off_y, float off_z,
+                         ssize_t indices[8], float weights[8])
 {
 	// Get voxel size
 	const float vx = length_x / nx;
